Added additional items to planning db. Added to add ip address, and address blocks for planning db Added table for passwords to planning db
